made some changes

+++ 0x15-file_io/0-read_textfile.c
@@ -8,7 +8,7 @@
  * Return: If filename is NULL return 0
  */
 
-ssize_t read_textfile(const char *filename, int letters)
+ssize_t read_textfile(const char *filename, size_t letters)
 {
     if (filename == NULL)
         return 0;
